data-explorer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-explorer"
version = "0.1.0"
description = "A LangChain and LangGraph-based agent for exploring NASA AppEEARS data"
authors = [
    { name = "NASA Space Apps Chicago Team" }
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-mcp-adapters",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0.0",
    "gradio>=4.19.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "isort>=5.13.2",
]

[project.scripts]
data-explorer = "data_explorer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/data_explorer"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310" 